feat: add id and description for the VIP 323
