+++ simulations/vip-323/bscmainnet.ts
@@ -4,7 +4,7 @@ import { ethers } from "hardhat";
 
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import vip322, {
+import vip323, {
   COMMUNITY_WALLET,
   ETH,
   ETH_AMOUNT_WALLET,
@@ -12,7 +12,7 @@ import vip322, {
   XVS_AMOUNT,
   XVS_AMOUNT_WALLET,
   XVS_BRIDGE_SRC,
-} from "../../vips/vip-322/bscmainnet";
+} from "../../vips/vip-323/bscmainnet";
 import ERC20_ABI from "./abi/ERC20.json";
 import XVS_BRIDGE_ABI from "./abi/XVSProxyOFTSrc.json";
 
@@ -36,7 +36,7 @@ forking(39546351, () => {
     oldEthBalWallet = await eth.balanceOf(COMMUNITY_WALLET);
   });
 
-  testVip("VIP-322 Send XVS to Dest Chain", vip322(), {
+  testVip("VIP-323 Send XVS to Dest Chain", vip323(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [XVS_BRIDGE_ABI], ["SendToChain"], [1]);
     },
